feat: Update app configuration for Docker container support
