+++ tests/index.test.ts
@@ -1,17 +1,11 @@
 import request from 'supertest';
-const { afterAll } = require('@jest/globals');
 
-import app from '../src/index';
-import { server } from '../src/index';
+import app from '../src/app';
 
 describe('GET /', () => {
   it('should return "Hello World!"', async () => {
-    const response = await request(app).get("/");
+    const response = await request(app).get('/');
     expect(response.status).toBe(200);
     expect(response.body).toEqual({ message: 'Hello World!' });
   });
-});
-
-afterAll(() => {
-    server.close();
 });
